// src/store/tasks/types.ts
export enum TasksActionsType {
  ADD_TASK = 'ADD_TASK',
  CHANGE_TASK_DUE_DATE = 'CHANGE_TASK_DUE_DATE',
  DELETE_TASK = 'DELETE_TASK',
  MARK_TASK_AS_DONE = 'MARK_TASK_AS_DONE',
  SET_IS_FROM_SECTION = 'SET_IS_FROM_SECTION',
  MARK_SUBTASK_AS_DONE = 'MARK_SUBTASK_AS_DONE',
  MARK_SUBTASK_AS_UNDONE = 'MARK_SUBTASK_AS_UNDONE',
  ADD_SUBTASKS_TO_TASK = 'ADD_SUBTASKS_TO_TASK',
  DELETE_SUBTASKS = 'DELETE_SUBTASKS',
  SET_SELECTED_TASK = 'SET_SELECTED_TASK',
}